add validate_schema api module
