+++ server/src/api.rs
@@ -1,2 +1,5 @@
 pub mod chat_completions;
+pub mod validate_schema;
+
 pub use chat_completions::*;
+pub use validate_schema::*;
